+++ app/controllers/legal_panel_for_government/rm6360/admin/sessions_controller.rb
@@ -1,0 +1,9 @@
+module LegalPanelForGovernment
+  module RM6360
+    module Admin
+      class SessionsController < Base::SessionsController
+        include LegalPanelForGovernment::Admin::FrameworkStatusConcern
+      end
+    end
+  end
+end