+++ app/models/legal_panel_for_government/rm6360/journey/suppliers.rb
@@ -1,0 +1,7 @@
+module LegalPanelForGovernment
+  module RM6360
+    class Journey::Suppliers
+      include Steppable
+    end
+  end
+end